# Handle surrogate pairs during scanning