--- a/lib/src/utils.dart
+++ b/lib/src/utils.dart
@@ -44,52 +44,8 @@ YamlWarningCallback yamlWarningCallback = (message, [SourceSpan? span]) {
   print(message);
 };
 
-// The following utility functions are copied from the string_scanner package.
-//
-// See https://en.wikipedia.org/wiki/UTF-16#Code_points_from_U+010000_to_U+10FFFF
-// for documentation on how UTF-16 encoding works and definitions of various
-// related terms.
+/// Whether [codeUnit] is a UTF-16 high surrogate.
+bool isHighSurrogate(int codeUnit) => codeUnit >>> 10 == 0x36;
 
-/// The inclusive lower bound of Unicode's supplementary plane.
-const _supplementaryPlaneLowerBound = 0x10000;
-
-/// The inclusive upper bound of Unicode's supplementary plane.
-const _supplementaryPlaneUpperBound = 0x10FFFF;
-
-/// The inclusive lower bound of the UTF-16 high surrogate block.
-const _highSurrogateLowerBound = 0xD800;
-
-/// The inclusive lower bound of the UTF-16 low surrogate block.
-const _lowSurrogateLowerBound = 0xDC00;
-
-/// The number of low bits in each code unit of a surrogate pair that goes into
-/// determining which code point it encodes.
-const _surrogateBits = 10;
-
-/// A bit mask that covers the lower [_surrogateBits] of a code point, which can
-/// be used to extract the value of a surrogate or the low surrogate value of a
-/// code unit.
-const _surrogateValueMask = (1 << _surrogateBits) - 1;
-
-/// Returns whether [codePoint] is in the Unicode supplementary plane, and thus
-/// must be represented as a surrogate pair in UTF-16.
-bool inSupplementaryPlane(int codePoint) =>
-    codePoint >= _supplementaryPlaneLowerBound &&
-    codePoint <= _supplementaryPlaneUpperBound;
-
-/// Returns whether [codeUnit] is a UTF-16 high surrogate.
-bool isHighSurrogate(int codeUnit) =>
-    (codeUnit & ~_surrogateValueMask) == _highSurrogateLowerBound;
-
-/// Returns whether [codeUnit] is a UTF-16 low surrogate.
-bool isLowSurrogate(int codeUnit) =>
-    (codeUnit >> _surrogateBits) == (_lowSurrogateLowerBound >> _surrogateBits);
-
-/// Converts a UTF-16 surrogate pair into the Unicode code unit it represents.
-int decodeSurrogatePair(int highSurrogate, int lowSurrogate) {
-  assert(isHighSurrogate(highSurrogate));
-  assert(isLowSurrogate(lowSurrogate));
-  return _supplementaryPlaneLowerBound +
-      (((highSurrogate & _surrogateValueMask) << _surrogateBits) |
-          (lowSurrogate & _surrogateValueMask));
-}
+/// Whether [codeUnit] is a UTF-16 low surrogate.
+bool isLowSurrogate(int codeUnit) => codeUnit >>> 10 == 0x37;