@natebosch 2024-03-27
@lrhn - should we be using `package:characters` or does this look sufficient?