@tamcy 2024-04-04
In the latest revision I changed back to use `readChar()` whenever possible. Also I have remove the need to decode the surrogate for further checking, because a valid surrogate pair itself should be of the range of `0x10000` to `0x10FFFF`, so just checking validity of the high and low surrogate is enough.